Organize project into core and ui folders
