+++ FusionFall-Mod/Core/Compression/7zip/Common/InBuffer.cs
