+++ FusionFall-Mod/Core/Compression/7zip/Common/OutBuffer.cs
