+++ FusionFall-Mod/Core/Compression/7zip/Compress/LZ/IMatchFinder.cs
