+++ FusionFall-Mod/Core/Compression/7zip/Compress/LZ/LzBinTree.cs
