+++ FusionFall-Mod/Core/Compression/7zip/Compress/LZ/LzInWindow.cs
