+++ FusionFall-Mod/Core/Compression/7zip/Compress/LZMA/LzmaBase.cs
