+++ FusionFall-Mod/Core/Compression/7zip/Compress/LZMA/LzmaEncoder.cs
