+++ FusionFall-Mod/Core/Compression/7zip/Compress/RangeCoder/RangeCoderBit.cs
