+++ FusionFall-Mod/Core/Compression/7zip/ICoder.cs
