+++ FusionFall-Mod/Core/EndianConverter.cs
