+++ FusionFall-Mod/Models/FileEntry.cs
