+++ FusionFall-Mod/Models/HeaderInfo.cs
